Move linear encoders to embedding.py
